+++ packages/terra-framework-docs/src/terra-dev-site/test/data-grid/data-grid/DefaultDataGrid.test.jsx
@@ -68,8 +68,6 @@ const DefaultDataGrid = () => {
         overflowColumns={cols}
         rows={rows}
         rowHeaderIndex={0}
-        // TODO: Remove columnHeaderHeight from example
-        columnHeaderHeight="50px"
         ariaLabel="Data Grid"
       />
     </React.Fragment>